Fix sizing error for bitfield
The mergeable field was define as an uint32_t, but MAX_SRC_NUM is now
40, so we need at least an uint64_t.

+++ backend/src/backend/gen_insn_selection.cpp
@@ -279,7 +279,7 @@ namespace gbe
     /*! Instruction that needs to be matched */
     const ir::Instruction &insn;
     /*! When sources have been overwritten, a child insn cannot be merged */
-    uint32_t mergeable:ir::Instruction::MAX_SRC_NUM;
+    uint64_t mergeable:ir::Instruction::MAX_SRC_NUM;
     /*! Number of children we have in the pattern */
     uint32_t childNum:7;
     /*! A root must be generated, no matter what */
